requirements.txt
# Google Cloud dependencies
google-cloud-aiplatform>=1.38.0
google-cloud-core>=2.3.0
google-auth>=2.23.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.7.0

# Async support
aiohttp>=3.8.0
asyncio-mqtt>=0.13.0

# Logging and monitoring
structlog>=23.1.0
rich>=13.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Development
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0 